// apps/api/src/modules/openai/dto/get-response-dtos/ResponseFunctionToolCallDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString } from 'class-validator';

export class ResponseFunctionToolCallDto {
  /** A JSON string of the arguments to pass to the function. */
  @ApiProperty({ description: `A JSON string of the arguments to pass to the function.`, type: 'string' })
  @IsString()
  arguments!: string;

  /** The unique ID of the function tool call generated by the model. */
  @ApiProperty({ description: `The unique ID of the function tool call generated by the model.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** The name of the function to run. */
  @ApiProperty({ description: `The name of the function to run.`, type: 'string' })
  @IsString()
  name!: string;

  /** The type of the function tool call. Always `function_call`. */
  @ApiProperty({ description: `The type of the function tool call. Always \`function_call\`.`, example: 'function_call' })
  @Equals('function_call')
  type!: 'function_call';

  /** The unique ID of the function tool call. */
  @ApiProperty({ required: false, description: `The unique ID of the function tool call.`, type: 'string' })
  @IsOptional()
  @IsString()
  id?: string;

  /** The namespace of the function to run. */
  @ApiProperty({ required: false, description: `The namespace of the function to run.`, type: 'string' })
  @IsOptional()
  @IsString()
  namespace?: string;

  /** The status of the item. One of `in_progress`, `completed`, or `incomplete`.
Populated when items are returned via API. */
  @ApiProperty({ required: false, description: `The status of the item. One of \`in_progress\`, \`completed\`, or \`incomplete\`.
Populated when items are returned via API.`, enum: ['in_progress', 'completed', 'incomplete'] })
  @IsOptional()
  @IsIn(['in_progress', 'completed', 'incomplete'])
  status?: 'in_progress' | 'completed' | 'incomplete';
}