// apps/api/src/modules/openai/dto/get-response-dtos/ResponseCodeInterpreterToolCallDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsIn, IsString } from 'class-validator';

import { LogsDto } from './LogsDto';
import { ImageDto } from './ImageDto';

export class ResponseCodeInterpreterToolCallDto {
  /** The unique ID of the code interpreter tool call. */
  @ApiProperty({ description: `The unique ID of the code interpreter tool call.`, type: 'string' })
  @IsString()
  id!: string;

  /** The code to run, or null if not available. */
  @ApiProperty({ description: `The code to run, or null if not available.` })
  code!: null | string;

  /** The ID of the container used to run the code. */
  @ApiProperty({ description: `The ID of the container used to run the code.`, type: 'string' })
  @IsString()
  container_id!: string;

  /** The outputs generated by the code interpreter, such as logs or images. Can be
null if no outputs are available. */
  @ApiProperty({ description: `The outputs generated by the code interpreter, such as logs or images. Can be
null if no outputs are available.` })
  outputs!: null | LogsDto | ImageDto[];

  /** The status of the code interpreter tool call. Valid values are `in_progress`,
`completed`, `incomplete`, `interpreting`, and `failed`. */
  @ApiProperty({ description: `The status of the code interpreter tool call. Valid values are \`in_progress\`,
\`completed\`, \`incomplete\`, \`interpreting\`, and \`failed\`.`, enum: ['in_progress', 'completed', 'incomplete', 'failed', 'interpreting'] })
  @IsIn(['in_progress', 'completed', 'incomplete', 'failed', 'interpreting'])
  status!: 'in_progress' | 'completed' | 'incomplete' | 'failed' | 'interpreting';

  /** The type of the code interpreter tool call. Always `code_interpreter_call`. */
  @ApiProperty({ description: `The type of the code interpreter tool call. Always \`code_interpreter_call\`.`, example: 'code_interpreter_call' })
  @Equals('code_interpreter_call')
  type!: 'code_interpreter_call';
}