// apps/api/src/modules/openai/dto/get-response-dtos/ResponseCustomToolCallDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsOptional, IsString } from 'class-validator';

export class ResponseCustomToolCallDto {
  /** An identifier used to map this custom tool call to a tool call output. */
  @ApiProperty({ description: `An identifier used to map this custom tool call to a tool call output.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** The input for the custom tool call generated by the model. */
  @ApiProperty({ description: `The input for the custom tool call generated by the model.`, type: 'string' })
  @IsString()
  input!: string;

  /** The name of the custom tool being called. */
  @ApiProperty({ description: `The name of the custom tool being called.`, type: 'string' })
  @IsString()
  name!: string;

  /** The type of the custom tool call. Always `custom_tool_call`. */
  @ApiProperty({ description: `The type of the custom tool call. Always \`custom_tool_call\`.`, example: 'custom_tool_call' })
  @Equals('custom_tool_call')
  type!: 'custom_tool_call';

  /** The unique ID of the custom tool call in the OpenAI platform. */
  @ApiProperty({ required: false, description: `The unique ID of the custom tool call in the OpenAI platform.`, type: 'string' })
  @IsOptional()
  @IsString()
  id?: string;

  /** The namespace of the custom tool being called. */
  @ApiProperty({ required: false, description: `The namespace of the custom tool being called.`, type: 'string' })
  @IsOptional()
  @IsString()
  namespace?: string;
}