// apps/api/src/modules/openai/dto/get-response-dtos/ResponseFunctionToolCallOutputItemDto.ts
import { ApiExtraModels, ApiProperty, getSchemaPath } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString } from 'class-validator';

import { ResponseInputTextDto } from './ResponseInputTextDto';
import { ResponseInputImageDto } from './ResponseInputImageDto';
import { ResponseInputFileDto } from './ResponseInputFileDto';

@ApiExtraModels(
  ResponseInputTextDto,
  ResponseInputImageDto,
  ResponseInputFileDto,
)
export class ResponseFunctionToolCallOutputItemDto {
  /** The unique ID of the function call tool output. */
  @ApiProperty({
    description: `The unique ID of the function call tool output.`,
    type: 'string',
  })
  @IsString()
  id!: string;

  /** The unique ID of the function tool call generated by the model. */
  @ApiProperty({
    description: `The unique ID of the function tool call generated by the model.`,
    type: 'string',
  })
  @IsString()
  call_id!: string;

  /**
   * The output from the function call generated by your code. Can be a string or an
   * list of output content.
   */
  @ApiProperty({
    description: `The output from the function call generated by your code. Can be a string or an
  list of output content.`,
    isArray: true,
    oneOf: [
      { type: 'string' },
      { $ref: getSchemaPath(ResponseInputTextDto) },
      { $ref: getSchemaPath(ResponseInputImageDto) },
      { $ref: getSchemaPath(ResponseInputFileDto) },
    ],
  })
  output!: string | (ResponseInputTextDto | ResponseInputImageDto | ResponseInputFileDto)[];

  /**
   * The status of the item. One of `in_progress`, `completed`, or `incomplete`.
   * Populated when items are returned via API.
   */
  @ApiProperty({
    description: `The status of the item. One of \`in_progress\`, \`completed\`, or \`incomplete\`.
  Populated when items are returned via API.`,
    enum: ['in_progress', 'completed', 'incomplete'],
  })
  @IsIn(['in_progress', 'completed', 'incomplete'])
  status!: 'in_progress' | 'completed' | 'incomplete';

  /** The type of the function tool call output. Always `function_call_output`. */
  @ApiProperty({
    description: `The type of the function tool call output. Always \`function_call_output\`.`,
    type: 'string',
    enum: ['function_call_output'],
  })
  @Equals('function_call_output')
  type!: 'function_call_output';

  /** The identifier of the actor that created the item. */
  @ApiProperty({
    required: false,
    description: `The identifier of the actor that created the item.`,
    type: 'string',
  })
  @IsOptional()
  @IsString()
  created_by?: string;
}