// apps/api/src/modules/openai/dto/get-response-dtos/ResponseFunctionShellToolCallDto.ts
import { ApiExtraModels, ApiProperty, getSchemaPath } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString, ValidateNested } from 'class-validator';
import { Type } from 'class-transformer';

import { ActionDto } from './ActionDto';
import { ResponseLocalEnvironmentDto } from './ResponseLocalEnvironmentDto';
import { ResponseContainerReferenceDto } from './ResponseContainerReferenceDto';

@ApiExtraModels(
  ResponseLocalEnvironmentDto,
  ResponseContainerReferenceDto,
)
export class ResponseFunctionShellToolCallDto {
  /**
   * The unique ID of the shell tool call. Populated when this item is returned via
   * API.
   */
  @ApiProperty({
    description: `The unique ID of the shell tool call. Populated when this item is returned via
  API.`,
    type: 'string',
  })
  @IsString()
  id!: string;

  /** The shell commands and limits that describe how to run the tool call. */
  @ApiProperty({
    description: `The shell commands and limits that describe how to run the tool call.`,
    type: () => ActionDto,
  })
  @ValidateNested()
  @Type(() => ActionDto)
  action!: ActionDto;

  /** The unique ID of the shell tool call generated by the model. */
  @ApiProperty({
    description: `The unique ID of the shell tool call generated by the model.`,
    type: 'string',
  })
  @IsString()
  call_id!: string;

  /** Represents the use of a local environment to perform shell actions. */
  @ApiProperty({
    description: `Represents the use of a local environment to perform shell actions.`,
    oneOf: [
      { $ref: getSchemaPath(ResponseLocalEnvironmentDto) },
      { $ref: getSchemaPath(ResponseContainerReferenceDto) },
    ],
  })
  environment!: null | ResponseLocalEnvironmentDto | ResponseContainerReferenceDto;

  /**
   * The status of the shell call. One of `in_progress`, `completed`, or
   * `incomplete`.
   */
  @ApiProperty({
    description: `The status of the shell call. One of \`in_progress\`, \`completed\`, or
  \`incomplete\`.`,
    enum: ['in_progress', 'completed', 'incomplete'],
  })
  @IsIn(['in_progress', 'completed', 'incomplete'])
  status!: 'in_progress' | 'completed' | 'incomplete';

  /** The type of the item. Always `shell_call`. */
  @ApiProperty({
    description: `The type of the item. Always \`shell_call\`.`,
    type: 'string',
    enum: ['shell_call'],
  })
  @Equals('shell_call')
  type!: 'shell_call';

  /** The ID of the entity that created this tool call. */
  @ApiProperty({
    required: false,
    description: `The ID of the entity that created this tool call.`,
    type: 'string',
  })
  @IsOptional()
  @IsString()
  created_by?: string;
}