// apps/api/src/modules/openai/dto/get-response-dtos/ResponseApplyPatchToolCallOutputDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString } from 'class-validator';

export class ResponseApplyPatchToolCallOutputDto {
  /**
   * The unique ID of the apply patch tool call output. Populated when this item is
   * returned via API.
   */
  @ApiProperty({
    description: `The unique ID of the apply patch tool call output. Populated when this item is
  returned via API.`,
    type: 'string',
  })
  @IsString()
  id!: string;

  /** The unique ID of the apply patch tool call generated by the model. */
  @ApiProperty({
    description: `The unique ID of the apply patch tool call generated by the model.`,
    type: 'string',
  })
  @IsString()
  call_id!: string;

  /** The status of the apply patch tool call output. One of `completed` or `failed`. */
  @ApiProperty({
    description: `The status of the apply patch tool call output. One of \`completed\` or \`failed\`.`,
    enum: ['completed', 'failed'],
  })
  @IsIn(['completed', 'failed'])
  status!: 'completed' | 'failed';

  /** The type of the item. Always `apply_patch_call_output`. */
  @ApiProperty({
    description: `The type of the item. Always \`apply_patch_call_output\`.`,
    type: 'string',
    enum: ['apply_patch_call_output'],
  })
  @Equals('apply_patch_call_output')
  type!: 'apply_patch_call_output';

  /** The ID of the entity that created this tool call output. */
  @ApiProperty({
    required: false,
    description: `The ID of the entity that created this tool call output.`,
    type: 'string',
  })
  @IsOptional()
  @IsString()
  created_by?: string;

  /** Optional textual output returned by the apply patch tool. */
  @ApiProperty({
    required: false,
    description: `Optional textual output returned by the apply patch tool.`,
    type: 'string',
  })
  @IsOptional()
  output?: null | string;
}