// apps/api/src/modules/openai/dto/get-response-dtos/ResponseCustomToolCallOutputItemDto.ts
import { ApiExtraModels, ApiProperty, getSchemaPath } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString } from 'class-validator';

import { ResponseInputTextDto } from './ResponseInputTextDto';
import { ResponseInputImageDto } from './ResponseInputImageDto';
import { ResponseInputFileDto } from './ResponseInputFileDto';

@ApiExtraModels(
  ResponseInputTextDto,
  ResponseInputImageDto,
  ResponseInputFileDto,
)
export class ResponseCustomToolCallOutputItemDto {
  /** The call ID, used to map this custom tool call output to a custom tool call. */
  @ApiProperty({
    description: `The call ID, used to map this custom tool call output to a custom tool call.`,
    type: 'string',
  })
  @IsString()
  call_id!: string;

  /**
   * The output from the custom tool call generated by your code. Can be a string or
   * an list of output content.
   */
  @ApiProperty({
    description: `The output from the custom tool call generated by your code. Can be a string or
  an list of output content.`,
    isArray: true,
    oneOf: [
      { type: 'string' },
      { $ref: getSchemaPath(ResponseInputTextDto) },
      { $ref: getSchemaPath(ResponseInputImageDto) },
      { $ref: getSchemaPath(ResponseInputFileDto) },
    ],
  })
  output!: string | (ResponseInputTextDto | ResponseInputImageDto | ResponseInputFileDto)[];

  /** The type of the custom tool call output. Always `custom_tool_call_output`. */
  @ApiProperty({
    description: `The type of the custom tool call output. Always \`custom_tool_call_output\`.`,
    type: 'string',
    enum: ['custom_tool_call_output'],
  })
  @Equals('custom_tool_call_output')
  type!: 'custom_tool_call_output';

  /** The unique ID of the custom tool call output item. */
  @ApiProperty({
    description: `The unique ID of the custom tool call output item.`,
    type: 'string',
  })
  @IsString()
  id!: string;

  /**
   * The status of the item. One of `in_progress`, `completed`, or `incomplete`.
   * Populated when items are returned via API.
   */
  @ApiProperty({
    description: `The status of the item. One of \`in_progress\`, \`completed\`, or \`incomplete\`.
  Populated when items are returned via API.`,
    enum: ['in_progress', 'completed', 'incomplete'],
  })
  @IsIn(['in_progress', 'completed', 'incomplete'])
  status!: 'in_progress' | 'completed' | 'incomplete';

  /** The identifier of the actor that created the item. */
  @ApiProperty({
    required: false,
    description: `The identifier of the actor that created the item.`,
    type: 'string',
  })
  @IsOptional()
  @IsString()
  created_by?: string;
}